#31: moved findChrome and exported utility functions
