+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "markugen",
-  "version": "1.1.1",
-  "description": "Markdown to HTML static site generation tool",
+  "version": "1.1.2",
+  "description": "Markdown to HTML/PDF static site generation tool",
   "repository": {
     "type": "git",
     "url": "https://github.com/falkorclark/markugen.git"
