+++ readme.md
@@ -44,6 +44,7 @@ of the major features are listed below:
 ## Major Features
 
 * Markdown to HTML
+* Markdown to PDF
 * Website generation
 * Static HTML generation (no need for a server)
 * Reactive website sitemap and Table of Contents
