+++ src/markugen.ts
@@ -6,12 +6,14 @@ import os from 'node:os';
 import Generator from './generator';
 import { version, name, homepage } from '../package.json';
 import { Options, Themes } from './options';
-import { findChrome, timeFormat } from './utils';
+import { timeFormat } from './utils';
 import { defaultThemes } from './themes';
 import Preprocessor from './preprocessor';
 import IMarkugen from './imarkugen';
+import { spawnSync } from 'node:child_process';
 
 export * from './options';
+export * from './utils';
 
 export interface OutputLabel 
 {
@@ -357,6 +359,77 @@ export default class Markugen
       if (path.indexOf(exclude) !== -1) 
         return true;
     return false;
+  }
+  
+  /**
+   * Attempts to locate an executable for Google Chrome
+   * @returns the path to chrome if found, else returns undefined
+   */
+  public static findChrome():string|undefined
+  {
+    switch(process.platform)
+    {
+      case 'win32': return Markugen.findChromeWindows();
+      case 'darwin': return Markugen.findChromeMac();
+      default: return Markugen.findChromeLinux();
+    }
+  }
+
+  /**
+   * Attempts to locate an executable for Google Chrome on Windows
+   * @returns the path to chrome if found, else returns undefined
+   */
+  public static findChromeWindows():string|undefined
+  {
+    const suffix = '\\Google\\Chrome\\Application\\chrome.exe';
+    const prefixes = [
+      process.env.LOCALAPPDATA,
+      process.env.PROGRAMFILES,
+      process.env['PROGRAMFILES(X86)']
+    ];
+    for (const prefix of prefixes)
+    {
+      if (prefix)
+      {
+        const test = path.join(prefix, suffix);
+        if (fs.existsSync(test)) return test;
+      }
+    }
+    return undefined;
+  }
+
+  /**
+   * Attempts to locate an executable for Google Chrome on Mac
+   * @returns the path to chrome if found, else returns undefined
+   */
+  public static findChromeMac():string|undefined
+  {
+    const toExec = '/Contents/MacOS/Google Chrome';
+    const regPath = '/Applications/Google Chrome.app' + toExec;
+    const altPath = path.join(os.homedir(), regPath);
+    const mdFindCmd = 'mdfind \'kMDItemDisplayName == "Google Chrome" && kMDItemKind == Application\'';
+
+    if (fs.existsSync(regPath)) return regPath;
+    if (fs.existsSync(altPath)) return altPath;
+
+    // try the md command last
+    const result = spawnSync(mdFindCmd);
+    if (result.status === 0 && result.stdout) 
+      return result.stdout.toString().trim() + toExec;
+
+    return undefined;
+  }
+
+  /**
+   * Attempts to locate an executable for Google Chrome on Linux
+   * @returns the path to chrome if found, else returns undefined
+   */
+  public static findChromeLinux():string|undefined
+  {
+    const result = spawnSync('which google-chrome');
+    if (result.status === 0 && result.stdout) 
+      return result.stdout.toString().trim();
+    return undefined;
   }
 
   /**
